+++ src/core/ast.ts
@@ -103,7 +103,7 @@ export class ConfidenceExpression extends Expression {
   }
 }
 
-export type BinaryOperator = '+' | '-' | '*' | '/' | '%' | '>' | '<' | '>=' | '<=' | '==' | '!=' | '&&' | '||' | '~>' | '~~' | '~??' | '~&&' | '~||' | '~||>' | '~@>' | '~+' | '~-' | '~*' | '~/' | '~==' | '~!=' | '~<' | '~>=' | '~<=' | '~.' | '.';
+export type BinaryOperator = '+' | '-' | '*' | '/' | '%' | '**' | '>' | '<' | '>=' | '<=' | '==' | '!=' | '&&' | '||' | '??' | '~>' | '~~' | '~??' | '~&&' | '~||' | '~||>' | '~@>' | '~+' | '~-' | '~*' | '~/' | '~==' | '~!=' | '~<' | '~>=' | '~<=' | '~.' | '.';
 
 export class BinaryExpression extends Expression {
   type: NodeType = 'BinaryExpression';